Temp. fix for beta 5 UI issues (Picker ant VStack with List)
